Has added directory controller & change root on server
Create the controllers directory on the server side. Making Controllers
is done to provide variables that will appear. Separation is done to make
error checking easier and more organized.

The renaming of the root api is done to change the initial initializer
to the actual program. Previously 'api/name' to 'api/porto'.